+++ src/main/java/nl/nn/testtool/MessageCapturer.java
@@ -17,6 +17,7 @@ package nl.nn.testtool;
 
 import java.io.OutputStream;
 import java.io.Writer;
+import java.util.function.Consumer;
 
 /**
  * Implement this interface to stream specific messages to a report.
@@ -49,13 +50,13 @@ public interface MessageCapturer {
 
 	/**
 	 * Get {@link StreamingType} of message. The {@link StreamingType} returned will determine whether
-	 * {@link #toWriter(Object, Writer)} or {@link #toOutputStream(Object, OutputStream)} will be called (or none of
+	 * {@link #toWriter(Object, Writer)} or {@link #toOutputStream(Object, OutputStream, Consumer)} will be called (or none of
 	 * them).
 	 * 
 	 * @param message  the message for which to return the {@link StreamingType}
 	 * @return         the {@link StreamingType} for the specified message
 	 */
-	StreamingType getStreamingType(Object message);
+	public StreamingType getStreamingType(Object message);
 
 	/**
 	 * This method will be called for messages of {@link StreamingType} {@link StreamingType#CHARACTER_STREAM} to write
@@ -64,29 +65,32 @@ public interface MessageCapturer {
 	 * closing and storing the report. The characters written to this writer will be passed as a string to
 	 * {@link Checkpoint#setMessage(String)}.
 	 * 
-	 * @param <T>           type of message
-	 * @param message       the checkpoint message object
-	 * @param writer        write the data of the checkpoint message object to this writer
-	 * @return              the message itself, or a wrapper around it that will be passed back to the caller. When
-	 *                      later the caller writes characters to the wrapper these characters can be captured by the
-	 *                      wrapper and copied to writer
+	 * @param <T>              type of message
+	 * @param message          the checkpoint message object
+	 * @param writer           write the data of the checkpoint message object to this writer
+	 * @return                 the message itself, or a wrapper around it that will be passed back to the caller. When
+	 *                         later the caller writes characters to the wrapper these characters can be captured by the
+	 *                         wrapper and copied to writer
 	 */
-	<T> T toWriter(T message, Writer writer);
+	public <T> T toWriter(T message, Writer writer);
 
 	/**
 	 * This method will be called for messages of {@link StreamingType} {@link StreamingType#BYTE_STREAM} to write a
 	 * representation of the message to the specified output stream. This can be done asynchronous in which case Ladybug
 	 * will continue to add checkpoints and when the report has finished will wait for the output stream to close before
 	 * closing and storing the report. The bytes written to this writer will be passed as a byte array to
-	 * {@link Checkpoint#setMessage(Object)}.
+	 * {@link Checkpoint#setMessage(Object)} using {@link MessageEncoder#toString(Object, String)} with the charset that
+	 * the application optionally notified the Ladybug of.
 	 * 
-	 * @param <T>           type of message
-	 * @param message       the checkpoint message object
-	 * @param outputStream  write the data of the checkpoint message object to this output stream
-	 * @return              the message itself, or a wrapper around it that will be passed back to the caller. When
-	 *                      later the caller writes bytes to the wrapper these bytes can be captured by the
-	 *                      wrapper and copied to outputStream
+	 * @param <T>              type of message
+	 * @param message          the checkpoint message object
+	 * @param outputStream     write the data of the checkpoint message object to this output stream
+	 * @param charsetNotifier  enables application to notify Ladybug of the charset that probably is the most proper to
+	 *                         render the outputStream to characters
+	 * @return                 the message itself, or a wrapper around it that will be passed back to the caller. When
+	 *                         later the caller writes bytes to the wrapper these bytes can be captured by the
+	 *                         wrapper and copied to outputStream
 	 */
-	<T> T toOutputStream(T message, OutputStream outputStream);
+	public <T> T toOutputStream(T message, OutputStream outputStream, Consumer<String> charsetNotifier);
 
 }